+++ swarm/CHANGELOG.md
@@ -59,6 +59,9 @@
 
 - Rename `ConnectionHandlerEvent::Custom` to `ConnectionHandlerEvent::NotifyBehaviour`. See [PR 3955].
 
+- Remove deprecated items. See [PR 3956].
+
+[PR 3292]: https://github.com/libp2p/rust-libp2p/pull/3292
 [PR 3605]: https://github.com/libp2p/rust-libp2p/pull/3605
 [PR 3651]: https://github.com/libp2p/rust-libp2p/pull/3651
 [PR 3715]: https://github.com/libp2p/rust-libp2p/pull/3715
@@ -73,7 +76,7 @@
 [PR 3919]: https://github.com/libp2p/rust-libp2p/pull/3919
 [PR 3927]: https://github.com/libp2p/rust-libp2p/pull/3927
 [PR 3955]: https://github.com/libp2p/rust-libp2p/pull/3955
-[PR 3292]: https://github.com/libp2p/rust-libp2p/pull/3292
+[PR 3956]: https://github.com/libp2p/rust-libp2p/pull/3956
 
 ## 0.42.2